+++ src/animus/core/intake/domain/entities/dtos/petition_document_dto.py
@@ -1,7 +1,8 @@
 from animus.core.shared.domain.decorators import dto
+from animus.core.shared.domain.structures import FilePath
 
 
 @dto
 class PetitionDocumentDto:
-    file_path: str
+    file_path: FilePath | str
     name: str