+++ src/animus/pipes/intake_pipe.py
@@ -10,17 +10,21 @@ from animus.core.intake.domain.errors import (
 from animus.core.intake.interfaces.analisyses_repository import AnalisysesRepository
 from animus.core.intake.interfaces.petitions_repository import PetitionsRepository
 from animus.core.shared.domain.errors import ForbiddenError
-from animus.core.shared.domain.structures import Id, Text
+from animus.core.shared.domain.structures import FilePath, Id
 from animus.pipes.auth_pipe import AuthPipe
 from animus.pipes.database_pipe import DatabasePipe
+from animus.validation.shared import IdSchema
 
 
 class IntakePipe:
     @staticmethod
-    def verify_analysis_by_account(
-        analysis_id: str,
-        account_id: Id,
-        analisyses_repository: AnalisysesRepository,
+    def verify_analysis_by_account_from_request(
+        analysis_id: IdSchema,
+        account_id: Annotated[Id, Depends(AuthPipe.get_account_id_from_request)],
+        analisyses_repository: Annotated[
+            AnalisysesRepository,
+            Depends(DatabasePipe.get_analisyses_repository_from_request),
+        ],
     ) -> Analysis:
         analysis = analisyses_repository.find_by_id(Id.create(analysis_id))
         if analysis is None:
@@ -33,7 +37,7 @@ class IntakePipe:
 
     @staticmethod
     def verify_petition_document_path_by_account(
-        petition_id: str,
+        petition_id: IdSchema,
         account_id: Annotated[Id, Depends(AuthPipe.get_account_id_from_request)],
         petitions_repository: Annotated[
             PetitionsRepository,
@@ -43,7 +47,7 @@ class IntakePipe:
             AnalisysesRepository,
             Depends(DatabasePipe.get_analisyses_repository_from_request),
         ],
-    ) -> Text:
+    ) -> FilePath:
         petition = petitions_repository.find_by_id(Id.create(petition_id))
         if petition is None:
             raise PetitionNotFoundError